Add a double version of FInterpTo. #jira UE-88354 #rb steve.robb #preflight 6063101db4299e0001c1ecaf
[CL 15875719 by Patrick Boutot in ue5-main branch]

+++ Engine/Source/Runtime/Slate/Private/Widgets/Views/STableViewBase.cpp
@@ -257,7 +257,7 @@ void STableViewBase::Tick( const FGeometry& AllottedGeometry, const double InCur
 
 			if (bEnableAnimatedScrolling)
 			{
-				CurrentScrollOffset = FMath::FInterpTo(CurrentScrollOffset, TargetScrollOffset, InDeltaTime, 12.f);
+				CurrentScrollOffset = FMath::FInterpTo(CurrentScrollOffset, TargetScrollOffset, (double)InDeltaTime, 12.0);
 				if (FMath::IsNearlyEqual(CurrentScrollOffset, TargetScrollOffset, 0.01))
 				{
 					CurrentScrollOffset = TargetScrollOffset;
